TightBindingToolkit.Chern

TightBindingToolkit.Chern.ChernNumber — Method

ChernNumber(Ham::Hamiltonian, subset::Vector{Int64}) --> Float64

Function to get Chern numbers given a Hamiltonian and a subset of bands

TightBindingToolkit.Chern.FieldStrength — Method

FieldStrength(Links::Tuple{Matrix{ComplexF64}, Matrix{ComplexF64}}) --> Matrix{ComplexF64}

Function to calculate the product of the links over each plaquette on the BZ grid. This is the generalized Bery curvature for multiple degenerate bands.

TightBindingToolkit.Chern.FindLinks — Method

FindLinks(Ham::Hamiltonian, subset::Vector{Int64}) --> Tuple{Matrix{ComplexF64}, Matrix{ComplexF64}}

Function to get the linking matrices on each neighbouring point in the BZ. On a bond connecting ki and kj, the linking matrix U is defined such that U[m, n] = <v^m[ki]|v^n[kj]> where states[kj[1], kj[2], :, m] v^m[kj], the mth eigenstate at momentum kj.