TightBindingToolkit.Hamiltonian

TightBindingToolkit.Hamiltonian — Type

Hamiltonian is a data type representing a general momentum-space Hamiltonian corresponding to the given UnitCell and BZ.

Attributes

H :: Matrix{Matrix{ComplexF64}}: A matrix (corresponding to the matrix of k-points in BZ) of Hamiltonian matrices.

bands :: Matrix{Vector{Float64}}: A matrix (corresponding to the matrix of k-points in BZ) of band spectrums.

states :: Matrix{Matrix{ComplexF64}}: A matrix (corresponding to the matrix of k-points in BZ) of band wavefunctions.

bandwidth :: Tuple{Float64, Float64} : the tuple of minimum and maximum energies in the band structure.

Initialize this structure using

Hamiltonian(uc::UnitCell, bz::BZ)

TightBindingToolkit.DOS — Method

DOS(Omega::Float64, Ham::Hamiltonian; till_band::Int64=length(Ham.bands[1, 1]), spread::Float64=1e-3) --> Float64
DOS(Omegas::Vector{Float64}, Ham::Hamiltonian; till_band::Int64=length(Ham.bands[1, 1]), spread::Float64=1e-3) --> Vector{Float64}

Calculate the Density of State correspondingto the given energies in Omegas, for the lowest bands upto till_band. The calculation is done at a finite spread of the delta-function sum.

TightBindingToolkit.DiagonalizeHamiltonian! — Method

DiagonalizeHamiltonian!(Ham::Hamiltonian)

Diagonalize the Hamiltonian at all momentum points in the BZ.

TightBindingToolkit.FillHamiltonian — Method

FillHamiltonian(uc::UnitCell, k::Vector{Float64} ; SpinMatrices::Vector{Matrix{ComplexF64}} = SpinMats((uc.localDim-1)//2)) :: Matrix{ComplexF64}

Returns the Hamiltonian at momentum point k, corresponding to the bonds present in UnitCell.