TightBindingToolkit.Lattice

TightBindingToolkit.LatticeStruct.Lattice — Type

Lattice{T} is a data type representing a general real-space lattice constructed out of a UnitCell{T}.

Attributes

uc :: UnitCell{T}: Unit Cell of the lattice.

size :: Vector{Int64}: size of the lattice in units of the UnitCell primitives.

length :: Int64: total number of sites.

sites :: Dict{ Tuple{Int64, Vector{Int64}}, Int64} : a dictionary with key = (sublattice, offset), and the corresponding value being the site number on the lattice.

positions :: Vector{ Vector{Float64}}: real-space positions of all the lattice sites.

fields :: Vector{ Vector{Float64}}: the fields on all the lattice sites.

BondSites :: Matrix{Int64}: a matrix with lattice.length rows such that BondSites[s, i] is the site number of the ith neighbour of site-s on the lattice.

BondDists :: Matrix{Float64}: a matrix with lattice.length rows such that BondDists[s, i] is the distance to the ith neighbour of site-s on the lattice.

BondLabels :: Matrix{String}: a matrix with lattice.length rows such that BondLabels[s, i] is the label of the bond to the ith neighbour of site-s on the lattice.

BondMats :: Matrix{Array{ComplexF64, T}}: a matrix with lattice.length rows such that BondMats[s, i] is the Array{ComplexF64, T} of the bond connecting to the ith neighbour of site-s on the lattice.

Initialize this structure using

Lattice( uc::UnitCell{T}, size::Vector{Int64} ; null_dist::Float64 = -1.0, null_label::String = "-")

where null_dist and null_label are used for bonds which are not allowed due to given boundary conditions, but still tracked in the code.

TightBindingToolkit.LatticeStruct.ApplyBCToSite — Method

ApplyBCToSite(site::Vector{Int64}, L::Vector{Int64}, BC::Vector{ComplexF64}) -->  Vector{Int64}

Returns the effective real-space offset of site given a lattice size L, and boundary conditions BC.

TightBindingToolkit.LatticeStruct.FillBonds! — Method

FillBonds!(lattice::Lattice{T} ; null_dist::Float64 = -1.0, null_label::String = "-" ) where {T}

Fills all the bond information in the lattice using the bonds in UnitCell.

TightBindingToolkit.LatticeStruct.FillLattice! — Method

FillLattice!( lattice::Lattice{T} ; null_dist::Float64 = -1.0, null_label::String = "-") where {T}

Wrapper function to fill both site and bond information in the lattice.

TightBindingToolkit.LatticeStruct.FillSites! — Method

FillSites!(lattice::Lattice{T})

Fills all the sites, positions, and fields of the lattice using information in the UnitCell.

TightBindingToolkit.LatticeStruct.GetBCPhase — Method

GetBCPhase(site::Vector{Int64}, L::Vector{Int64}, BC::Vector{ComplexF64}) --> ComplexF64

Returns the effective phase of site given a lattice size L, and boundary conditions BC.

TightBindingToolkit.DesignLattice.CreateLattice — Method

CreateLattice(uc::UnitCell{T}, param::Param{T}, size::Vector{Int64} , index::Int64=length(param.value) ; null_dist::Float64 = -1.0, null_label::String = "-") --> Lattice{T} where {T}
CreateLattice(uc::UnitCell{T}, params::Vector{Param{T}}, size::Vector{Int64} , indices::Vector{Int64}=length.(getproperty.(params, :value)) ; null_dist::Float64 = -1.0, null_label::String = "-") :: Lattice{T} where {T}

Creates a lattice using Param objetcs given in params.

TightBindingToolkit.DesignLattice.ModifyLattice! — Method

ModifyLattice!(lattice::Lattice{T}, param::Param{T}) where {T}
ModifyLattice!(lattice::Lattice{T}, params::Vector{Param{T}}) where {T}

Modifies a lattice when the Param objects given in params are modified.

TightBindingToolkit.DesignLattice.RemoveLatticeBonds! — Method

RemoveLatticeBonds!(lattice::Lattice{T} , label::String ; null_dist::Float64 = -1.0, null_label::String = "-") where {T}

Removes a lattice bond with the given label.

TightBindingToolkit.DesignLattice.ScaleLattice! — Method

ScaleLattice!(lattice::Lattice{T}, param::Param{T}) where {T}
ScaleLattice!(lattice::Lattice{T}, params::Vector{Param{T}}) where {T}

Scales a lattice bond assuming that the Param objects got their strengths modified.

TightBindingToolkit.DesignLattice.ScaleLatticeBonds! — Method

ScaleLatticeBonds!(lattice::Lattice{T} , label::String , scaling::Float64) where {T}

Scales a lattice bond with the given label and by the given scaling.