TightBindingToolkit.Model

TightBindingToolkit.Model — Type

Model is a data type representing a general Tight Binding system.

Attributes

uc :: UnitCell: the Unit cell of the lattice.

bz :: BZ: The discretized Brillouin Zone.

Ham :: Hamiltonian: the Hamiltonian at all momentum-points.

T :: Float64: the temperature of the system.

filling :: Float64: The filling of the system.

mu :: Float64: The chemical potential of the system.

stat :: Int64 : ±1 for bosons and fermions.

gap :: Float64 : the energy gap of excitations at the given filling.

Gk :: Matrix{Matrix{ComplexF64}} : A matrix (corresponding to the matrix of k-points in BZ) of Greens functions.

Initialize this structure using

Model(uc::UnitCell, bz::BZ, Ham::Hamiltonian ; T::Float64=0.0, filling::Float64=-1.0, mu::Float64=0.0, stat::Int64=-1)

You can either input a filling, or a chemical potential. The corresponding μ for a given filling, or filling for a given μ is automatically calculated.

TightBindingToolkit.SolveModel! — Method

SolveModel!(M::Model)

one-step function to find all the attributes in Model after it has been initialized.

TightBindingToolkit.dist — Method

dist(E ; T::Float64, mu::Float64=0.0, stat::Int64=-1)

Distribution function. stat=1 –> Bose-Einstein distribution, and stat=-1 –> Fermi distribution.

TightBindingToolkit.distDer — Method

distDer(E ; T::Float64, mu::Float64=0.0, stat::Int64=-1)

derivative of dist w.r.t the energy. stat=1 –> Bose-Einstein distribution, and stat=-1 –> Fermi distribution.

TightBindingToolkit.findFilling — Function

findFilling(bands::Vector{Float64}, mu::Float64, T::Float64, stat::Int64=-1) --> Float64

Find filling at given T=temperature and mu=chemical potential, for a given bands.

TightBindingToolkit.getCount — Method

getCount(Es::Vector{Float64}, mu::Float64, T::Float64, stat::Int64) --> Matrix{Float64}

Function to return a diagonal matrix whose entries are M[i, i] = θ(-(E^i(k)-μ)) ––> 1 if the energy is below the chemical potential, otherwise 0.

TightBindingToolkit.getFilling! — Method

getFilling!(M::Model)

Find filling for a given Model.

TightBindingToolkit.getGk! — Method

getGk!(M::Model)

Finding the equal-time Greens functions in momentum space of a Model.

TightBindingToolkit.getMu! — Function

getMu!(M::Model, tol::Float64=0.001)

Function to get chemical potential for a given Model, within a tolerance.