TightBindingToolkit.Model

TightBindingToolkit.PlotTB.plot_FS! — Method

plot_FS!(Ham::Hamiltonian , bz::BZ , Efermi::Vector{Float64} , band_index::Vector{Int64})--> Plots.plot()

Function to draw the fermi surface at Efermi for the given Hamiltonian on the given BZ.

TightBindingToolkit.PlotTB.plot_UnitCell! — Method

plot_UnitCell!(uc::UnitCell ; range::Int64 = 1, cmp::Symbol = :matter, plot_conjugate::Bool=true) --> Plots.plot()

Function to plot the UnitCell. range determines the range of UnitCells plotted in real-space. default is ±1. cmp determines the colorScheme of chosen to differentiate b.w different hopping types. plot_conjugate switches on whether the conjugate of the bonds in UnitCell are also plotted or not.

TightBindingToolkit.PlotTB.plot_band_contour! — Method

plot_band_contour!(Ham::Hamiltonian , bz::BZ , band_index::Int64) --> Plots.plot()

Function to draw equal energy contours of the bands in Hamiltonian, specifically for the band with the given band_index.

TightBindingToolkit.PlotTB.plot_band_structure! — Method

plot_band_structure!(M<:Union{Model, BdGModel}, path::Vector{Vector{Float64}},  band_index::Vector{Int64} = collect(1:length(M.Ham.bands[begin])) ; labels::Vector{} = repeat([""], length(path)), closed::Bool=true, nearest::Bool=true) --> Plots.plot()

Function to plot band structures of the Model along a path in the BZ determined by the given critical points. Can take in multiple bands into account ∈ band_index. labels are the Plot labels of the critical points.