TightBindingToolkit.UnitCell

TightBindingToolkit.Bond — Type

Bond{T<:Number} is a data type representing a general bond on a lattice.

Attributes

base::Int64: sub-lattice of the initial site on the bond.

base::Int64: sub-lattice of the final site on the bond.

offset::Vector{Int64}: the difference of the unit cells in which these sublattices belong to, in units of the lattice basis vectors.

mat::Matrix{T}: the matrtix describing this bond –> can be hopping for partons, or spin-exchange for spins.

dist::Float64: the distance b/w the two sites = length of bond.

label::String: some string label to mark the bond type.

TightBindingToolkit.UnitCell — Type

UnitCell is a data type representing a general unit cell of a lattice.

Attributes

primitives :: Vector{ Vector{ Float64 } }: primitive bases vectors of the lattice.

basis :: Vector{ Vector{ Float64 } }: positions of the sub-lattices.

bonds :: Vector{Bond}: the set of all bonds defining a lattice.

fields :: Vector{ Vector{Float64}} : the fields oneach basis site.

localDim :: Int64: Local Hilbert space dimension ( e.g. 3 for classical spins, 2 for spin-1/2 electrons ).

BC :: Vector{ Int64 }: boundary conditions.

Initialize this structure using

UnitCell( as::Vector{Vector{Float64}} , localDim::Int64)

TightBindingToolkit.getDistance — Function

getDistance(uc::UnitCell, base::Int64, target::Int64, offset::Vector{Int64}) --> Float64

get the distance between site at position (0, base) and (R, target), where R = offset, when written in units of the unit cell primitive vectors.

TightBindingToolkit.addBasisSite! — Function

addBasisSite!( uc::UnitCell , position::Vector{Float64} )
addBasisSite!( uc::UnitCell , position::Vector{Float64} , field::Vector{Float64} )

Add a sublattice to the UnitCell at the given real-space position, with an on-site field.

TightBindingToolkit.getAllOffsets — Function

getAllOffsets(OffsetRange::Int64, dim::Int64) --> Vector{Vector{Int64}}

Given a range, returns the set of all possible Bond offsets such that each element of the offset vector lies in [-OffsetRange, OffsetRange].

TightBindingToolkit.addAnisotropicBond! — Function

addAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Number , dist::Float64, label::String )
addAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Matrix{<:Number} , dist::Float64, label::String )

Add a bond with the given attributes to UnitCell. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond.

TightBindingToolkit.addIsotropicBonds! — Function

addIsotropicBonds!( uc::UnitCell , dist::Float64 , mats::Number , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)))
addIsotropicBonds!( uc::UnitCell , dist::Float64 , mats::Matrix{<:Number} , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)) )

Add a set of "isotropic" bonds, which are the same for each pair of sites at the given distance. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond. The input checkOffsetRange must be adjusted depending on the input distance. The optional input subs is meant for isotropic bonds when only a subset of sublattices are involved.

TightBindingToolkit.ModifyBonds! — Function

ModifyBonds!(uc::UnitCell, dist::Float64, newMat::Matrix{<:Number})
ModifyBonds!(uc::UnitCell, label::String, newMat::Matrix{<:Number})

Modify an existing bond in the UnitCell with the given label, or at a given distance=dist, to the given bond matrix.

TightBindingToolkit.ScaleBonds! — Function

ScaleBonds!(uc::UnitCell, dist::Float64, scale::Number)
ScaleBonds!(uc::UnitCell, label::String, scale::Number)

Scale the matrix of an existing bond in the UnitCell with the given label, or at a given distance=dist, by the given scaling factor.

TightBindingToolkit.RemoveBonds! — Function

RemoveBonds!(uc::UnitCell, dist::Float64)
ScaleBonds!(uc::UnitCell, label::String)

Remove an existing bond in the UnitCell with the given label, or at a given distance=dist.