TightBindingToolkit.BZ

TightBindingToolkit.BZ — Type

BZ is a data type representing a discretized Brillouin Zone in momentum space.

Attributes

basis :: Vector{ Vector{ Float64 } }: reciprocal lattice vectors of the Brillouin Zone.

gridSize :: Int64: The number of points along each dimension of the grid.

k1s :: Matrix{Float64}: the Monkhorst grid corresponding to k along b1.

k2s :: Matrix{Float64}: the Monkhorst grid corresponding to k along b2.

ks :: Matrix{Vector{Float64}}: The grid of momentum points [kx, ky].

HighSymPoints :: Dict: A dictionary containing the HIgh-Symmetry points Γ, K(2), and M(3).

shift :: Vector{Float64} : how shifted the grid is from its centre point at the Γ point.

Initialize this structure using

BZ(gridSize::Int64)

TightBindingToolkit.CombinedBZPath — Method

CombinedBZPath(bz::BZ, points::Vector{Vector{Float64}} ; nearest::Bool = false, closed::Bool = true) --> Vector{Vector{Float64}}

Returns a path in momentum-space of the discretized BZ which joins the given momentum points present in points as point[1] –> point[2] –> ... –> point[end] –> point[1]. The optional input nearest is the same as in GetQIndex, and closed determines if the path is a clsoed loop or not.

TightBindingToolkit.GetIndexPath — Method

GetIndexPath(start::Vector{Int64}, ending::Vector{Int64} ; exclusive::Bool=true) --> Vector{Vector{Int64}}

Returns a path in index-space of the discretized BZ which joins the two sets of indices start and ending. If the input exclusive is set to true, the returned path will NOT contain the ending point itself.

TightBindingToolkit.GetQIndex — Method

GetQIndex(Q::Vector{Float64}, bz::BZ ; nearest::Bool = false) --> Vector{Int64}

Returns the index in the discretized BZ of the momentum point corresponding to the fiven momentum Q. If the input nearest is set to true, will return the index of the momentum point on the grid closest to Q, if Q does not exist on the grid.

TightBindingToolkit.ReduceQ — Method

ReduceQ(Q::Vector{Float64}, bz::BZ) --> Vector{Float64}

Reduces a given momentum back to the range covered by the discretized Brillouin Zone.

TightBindingToolkit.fillBZ! — Function

fillBZ!(bz::BZ, uc::UnitCell, offsetRange::Int64=1 ; shift::Vector{Float64}=zeros(Float64, length(uc.primitives)))

Fills the BZ object with the relevant attributes, after it has been initialized as BZ(gridsize=N).

TightBindingToolkit.getBZPath — Method

getBZPath(bz::BZ, start::Vector{Float64}, ending::Vector{Float64} ; nearest::Bool = false, exclusive::Bool = true) --> Vector{Vector{Float64}}

Returns the actual path in momentum-space of the discretized BZ which joins the two momentums start and ending. The optional input nearest is the same as in GetQIndex, and exclusive is the same as in GetIndexPath.

TightBindingToolkit.getRLVs — Method

getRLVs( uc::UnitCell ) --> Vector{Vector{Float64}}

Returns the reciprocal lattice vectors corresponding to the given Unit Cell.