TightBindingToolkit.Hamiltonian

TightBindingToolkit.Hams.Hamiltonian — Type

Hamiltonian is a data type representing a general momentum-space Hamiltonian corresponding to the given UnitCell and BZ (or 2 Unit Cells if it is a BdG Hamiltonian).

Attributes

H :: Array{Matrix{ComplexF64}}: A Array (corresponding to the grid of k-points in BZ) of Hamiltonian matrices.

bands :: Array{Vector{Float64}}: A Array (corresponding to the grid of k-points in BZ) of band spectrums.

states :: Array{Matrix{ComplexF64}}: A Array (corresponding to the grid of k-points in BZ) of band wavefunctions.

bandwidth :: Tuple{Float64, Float64} : the tuple of minimum and maximum energies in the band structure.

is_BdG :: Bool : is the Hamiltonian a bdG hamiltonian or a pure hopping hamiltonian.

Initialize this structure using

Hamiltonian(uc::UnitCell, bz::BZ) --> Hopping Hamiltonian
Hamiltonian(uc_hop::UnitCell, uc_pair::UnitCell, bz::BZ) --> BdG Hamiltonian

TightBindingToolkit.Hams.DOS — Method

DOS(Omega::Float64, Ham::Hamiltonian; till_band::Int64=length(Ham.bands[1, 1]), spread::Float64=1e-3) --> Float64
DOS(Omegas::Vector{Float64}, Ham::Hamiltonian; till_band::Int64=length(Ham.bands[1, 1]), spread::Float64=1e-3) --> Vector{Float64}

Calculate the Density of State correspondingto the given energies in Omegas, for the lowest bands upto till_band. The calculation is done at a finite spread of the delta-function sum.

TightBindingToolkit.Hams.DiagonalizeHamiltonian! — Method

DiagonalizeHamiltonian!(Ham::Hamiltonian)

Diagonalize the Hamiltonian at all momentum points in the BZ. verbose is an optional argument to print when the Hamiltonian is diagonalized.

TightBindingToolkit.Hams.FillHamiltonian — Method

FullHamiltonian(uc::UnitCell, bz::BZ) --> Matrix{Matrix{ComplexF64}}

Returns the full Hamiltonian at all momentum points in BZ, corresponding to the bonds present in UnitCell.

TightBindingToolkit.Hams.FillHoppingHamiltonian — Method

FillHoppingHamiltonian(uc::UnitCell, k::Vector{Float64} ; SpinMatrices::Vector{Matrix{ComplexF64}}) --> Matrix{ComplexF64}

Returns the hopping Hamiltonian at momentum point k, corresponding to the bonds present in UnitCell.

TightBindingToolkit.Hams.FillPairingHamiltonian — Method

FillPairingHamiltonian(uc::UnitCell, k::Vector{Float64} ; SpinMatrices::Vector{Matrix{ComplexF64}}) --> Matrix{ComplexF64}

Returns the pairing Hamiltonian at momentum point k, corresponding to the bonds present in UnitCell.

TightBindingToolkit.Hams.ModifyHamiltonianField! — Method

ModifyHamiltonianField!(Ham::Hamiltonian, uc::UnitCell, newFields::Vector{Float64} ; dim::Int64=4, verbose::Bool=false)

Faster implementation of modifying ONLY the on-site field part of a Hamiltonian. newFields must be a vector of the same length as uc.basis. dim is an optional argument which determines which element of on-site field is being replaced.