TightBindingToolkit.Model

TightBindingToolkit.TBModel.Model — Type

Model is a data type representing a general Tight Binding system.

Attributes

uc :: UnitCell: the Unit cell of the lattice.

bz :: BZ: The discretized Brillouin Zone.

Ham :: Hamiltonian: the Hamiltonian at all momentum-points.

T :: Float64: the temperature of the system.

filling :: Float64: The filling of the system.

mu :: Float64: The chemical potential of the system.

stat :: Int64 : ±1 for bosons and fermions.

gap :: Float64 : the energy gap of excitations at the given filling.

Gk :: Array{Matrix{ComplexF64}} : An array (corresponding to the grid of k-points in BZ) of Greens functions.

Initialize this structure using

Model(uc::UnitCell, bz::BZ, Ham::Hamiltonian ; T::Float64=1e-3, filling::Float64=-1.0, mu::Float64=0.0, stat::Int64=-1)

You can either input a filling, or a chemical potential. The corresponding μ for a given filling, or filling for a given μ is automatically calculated.

TightBindingToolkit.TBModel.FindFilling — Function

FindFilling(bands::Vector{Float64}, mu::Float64, T::Float64, stat::Int64=-1) --> Float64

Find filling at given T=temperature and mu=chemical potential, for a given bands.

TightBindingToolkit.TBModel.FreeEnergy — Method

FreeEnergy(M::Model; F0::Float64 = 0.0) --> Float64

Calculate the free energy of the given Model.

TightBindingToolkit.TBModel.GetCount — Method

GetCount(Es::Vector{Float64}, mu::Float64, T::Float64, stat::Int64) --> Matrix{Float64}

Function to return a diagonal matrix whose entries are M[i, i] = θ(-(E^i(k)-μ)) ––> 1 if the energy is below the chemical potential, otherwise 0.

TightBindingToolkit.TBModel.GetFilling! — Method

GetFilling!(M::Model)

Find filling for a given Model.

TightBindingToolkit.TBModel.GetGk! — Method

GetGk!(M::Model)

Finding the equal-time Greens functions in momentum space of a Model.

TightBindingToolkit.TBModel.GetGr! — Method

GetGr!(M::Model)

Finding the equal-time Greens functions in real space of a Model.

TightBindingToolkit.TBModel.GetMu! — Method

GetMu!(M::Model; tol::Float64=0.001)

Function to get chemical potential for a given Model, within a tolerance.

TightBindingToolkit.TBModel.SolveModel! — Method

SolveModel!(M::Model)

one-step function to find all the attributes in Model after it has been initialized.