TightBindingToolkit.UnitCell

TightBindingToolkit.UCell.Bond — Type

Bond{T} is a data type representing a general bond on a lattice.

Attributes

base::Int64: sub-lattice of the initial site on the bond.

target::Int64: sub-lattice of the final site on the bond.

offset::Vector{Int64}: the difference of the unit cells in which these sublattices belong to, in units of the lattice basis vectors.

mat::Array{ComplexF64, T}: an array describing this bond –> can be hopping for partons, or spin-exchange for spins. The rank of the array, T, is determined by the specific model being looked at. Eg. rank = 2 for a free parton Hamiltonian.

dist::Float64: the distance b/w the two sites = length of bond.

label::String: some string label to mark the bond type.

TightBindingToolkit.UCell.UnitCell — Type

UnitCell{T} is a data type representing a general unit cell of a lattice.

Attributes

primitives :: Vector{ Vector{ Float64 } }: primitive bases vectors of the lattice.

basis :: Vector{ Vector{ Float64 } }: positions of the sub-lattices.

bonds :: Vector{Bond{T}}: the set of all bonds defining a lattice.

fields :: Vector{ Vector{Float64}} : the fields oneach basis site.

localDim :: Int64: Local Hilbert space dimension ( e.g. 3 for classical spins, 2 for spin-1/2 electrons ).

BC :: Vector{ ComplexF64 }: boundary conditions, in the form of [e^{ιθ_i}]. e.g. θ=0 for PBC, θ=π for APBC, and so on.

Initialize this structure using

UnitCell( as::Vector{Vector{Float64}} , localDim::Int64)
UnitCell( as::Vector{Vector{Float64}} , localDim::Int64, rank::Int64)

TightBindingToolkit.UCell.AddBasisSite! — Method

AddBasisSite!( uc::UnitCell , position::Vector{Float64} )
AddBasisSite!( uc::UnitCell , position::Vector{Float64} , field::Vector{Float64} )

Add a sublattice to the UnitCell at the given real-space position, with an on-site field.

TightBindingToolkit.UCell.BondRank — Method

BondRank(bond::Bond{T}) --> T
BondRank(bonds::Array{Bond{T}}) --> T

Function to return rank of a bond or a collection of bonds.

TightBindingToolkit.UCell.FlipBond — Method

FlipBond(bond::Bond{T}) --> Bond{T}

Function to flip a bond. Only works on bonds of rank = 2 right now.

TightBindingToolkit.UCell.GetDistance — Method

GetDistance(uc::UnitCell, base::Int64, target::Int64, offset::Vector{Int64}) --> Float64

get the distance between site at position (0, base) and (R, target), where R = offset, when written in units of the unit cell primitive vectors.

TightBindingToolkit.UCell.GetRealSpacePositions — Method

GetRealSpacePositions(uc::UnitCell{T} ; OffsetRange::Int64 = 2) --> Dict

Returns a dictionary whose keys are vectors in the cartesian coordinates (rounded off to accuracy digits), with values giving the corresponding sublattice and Unit Cell coordinate of a lattice site at that position.

TightBindingToolkit.UCell.IsSameBond — Method

IsSameBond( Bond1::Bond , Bond2::Bond ) --> Bool

Function to check if two bond objects are describing the same physical bond, just inverted!

TightBindingToolkit.UCell.IsSameUnitCell — Method

IsSameUnitCell(uc_1::UnitCell, uc_2::UnitCell) --> Bool

Function to check if two unit cell live on the same underlying lattice or not, i.e. have the same primitives, same sublattices, and same localDim.

TightBindingToolkit.DesignUCell.AddAnisotropicBond! — Method

AddAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Number , dist::Float64, label::String )
AddAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Matrix{<:Number} , dist::Float64, label::String )

Add a bond with the given attributes to UnitCell. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond.

TightBindingToolkit.DesignUCell.AddIsotropicBonds! — Method

AddIsotropicBonds!( uc::UnitCell{T} , dist::Float64 , mats::Number , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)))
AddIsotropicBonds!( uc::UnitCell{T} , dist::Float64 , mats::Array{<:Number, T} , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)) )

Add a set of "isotropic" bonds, which are the same for each pair of sites at the given distance. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond. The input checkOffsetRange must be adjusted depending on the input distance. The optional input subs is meant for isotropic bonds when only a subset of sublattices are involved.

TightBindingToolkit.DesignUCell.Lookup — Method

Lookup(uc::UnitCell) --> Dict

Returns a dictionary with keys = (base, target, offset) for bond ∈ UnitCell bond list, and the entry being the bond matrix. If there are multiple bonds with the same identifier, it adds them up.

TightBindingToolkit.DesignUCell.ModifyBonds! — Method

ModifyBonds!(uc::UnitCell, dist::Float64, newMat::Matrix{<:Number})
ModifyBonds!(uc::UnitCell, label::String, newMat::Matrix{<:Number})

Modify an existing bond in the UnitCell with the given label, or at a given distance=dist, to the given bond matrix.

TightBindingToolkit.DesignUCell.ModifyFields! — Method

ModifyFields!(uc::UnitCell, site::Int64, newField::Vector{Float64})
ModifyFields!(uc::UnitCell, newField::Vector{Vector{Float64}})

Modify the on-site fields in the UnitCell, either one at a time, or all of them.

TightBindingToolkit.DesignUCell.ModifyIsotropicFields! — Method

ModifyIsotropicFields!(uc::UnitCell, newField::Vector{Float64})
ModifyIsotropicFields!(uc::UnitCell, newField::Float64, dim::Int64)

Modify the on site field uniformly, on all sublattices. The optional argument dim is if you want to only modify one of the 4 elements of on-site fields (3 Zeeman and 1 chemical potential).

TightBindingToolkit.DesignUCell.RemoveBonds! — Method

RemoveBonds!(uc::UnitCell, dist::Float64)
ScaleBonds!(uc::UnitCell, label::String)

Remove an existing bond in the UnitCell with the given label, or at a given distance=dist.

TightBindingToolkit.DesignUCell.ScaleBonds! — Method

ScaleBonds!(uc::UnitCell, dist::Float64, scale::Number)
ScaleBonds!(uc::UnitCell, label::String, scale::Number)

Scale the matrix of an existing bond in the UnitCell with the given label, or at a given distance=dist, by the given scaling factor.

TightBindingToolkit.ExpandUCell.ChangePrimitives! — Method

ChangePrimitives!(uc::UnitCell{T}, newPrimitives::Vector{Vector{Float64}})

Changes the pirmitive vectors of the given UnitCell assuming the sublattices stay the same. Changes the bonds accordingly.

TightBindingToolkit.ExpandUCell.ExpandBonds! — Method

ExpandBonds!(ucOG::UnitCell{T}, ucNew::UnitCell{T} ; OffsetRange::Int64 = 2)

An in-place version of ExpandUnitCell which works with the new primitive vectors given in ucNew.

TightBindingToolkit.ExpandUCell.ExpandUnitCell — Method

ExpandUnitCell(ucOG::UnitCell{T}, scaling::Vector{Int64} ; OffsetRange::Int64 = 2)

Returns a UnitCell which is an integer multiple of the given UnitCell (the primitives are scaled by the given scalings along each direction). The bonds are properly redefined amongst the new sublattices and new primitive vectors.