TightBindingToolkit.UnitCell

TightBindingToolkit.UCell.Bond — Type

Bond{T<:Number} is a data type representing a general bond on a lattice.

Attributes

base::Int64: sub-lattice of the initial site on the bond.

base::Int64: sub-lattice of the final site on the bond.

offset::Vector{Int64}: the difference of the unit cells in which these sublattices belong to, in units of the lattice basis vectors.

mat::Matrix{T}: the matrtix describing this bond –> can be hopping for partons, or spin-exchange for spins.

dist::Float64: the distance b/w the two sites = length of bond.

label::String: some string label to mark the bond type.

TightBindingToolkit.UCell.UnitCell — Type

UnitCell is a data type representing a general unit cell of a lattice.

Attributes

primitives :: Vector{ Vector{ Float64 } }: primitive bases vectors of the lattice.

basis :: Vector{ Vector{ Float64 } }: positions of the sub-lattices.

bonds :: Vector{Bond}: the set of all bonds defining a lattice.

fields :: Vector{ Vector{Float64}} : the fields oneach basis site.

localDim :: Int64: Local Hilbert space dimension ( e.g. 3 for classical spins, 2 for spin-1/2 electrons ).

BC :: Vector{ ComplexF64 }: boundary conditions, in the form of [e^{ιθ_i}]. e.g. θ=0 for PBC, θ=π for APBC, and so on.

Initialize this structure using

UnitCell( as::Vector{Vector{Float64}} , localDim::Int64)

TightBindingToolkit.UCell.ModifyBonds! — Method

ModifyBonds!(uc::UnitCell, dist::Float64, newMat::Matrix{<:Number})
ModifyBonds!(uc::UnitCell, label::String, newMat::Matrix{<:Number})

Modify an existing bond in the UnitCell with the given label, or at a given distance=dist, to the given bond matrix.

TightBindingToolkit.UCell.ModifyFields! — Method

ModifyFields!(uc::UnitCell, site::Int64, newField::Vector{Float64})
ModifyFields!(uc::UnitCell, newField::Vector{Vector{Float64}})

Modify the on-site fields in the UnitCell, either one at a time, or all of them.

TightBindingToolkit.UCell.ModifyIsotropicFields! — Method

ModifyIsotropicFields!(uc::UnitCell, newField::Vector{Float64})
ModifyIsotropicFields!(uc::UnitCell, newField::Float64, dim::Int64)

Modify the on site field uniformly, on all sublattices. The optional argument dim is if you want to only modify one of the 4 elements of on-site fields (3 Zeeman and 1 chemical potential).

TightBindingToolkit.UCell.RemoveBonds! — Method

RemoveBonds!(uc::UnitCell, dist::Float64)
ScaleBonds!(uc::UnitCell, label::String)

Remove an existing bond in the UnitCell with the given label, or at a given distance=dist.

TightBindingToolkit.UCell.ScaleBonds! — Method

ScaleBonds!(uc::UnitCell, dist::Float64, scale::Number)
ScaleBonds!(uc::UnitCell, label::String, scale::Number)

Scale the matrix of an existing bond in the UnitCell with the given label, or at a given distance=dist, by the given scaling factor.

TightBindingToolkit.UCell.addAnisotropicBond! — Method

addAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Number , dist::Float64, label::String )
addAnisotropicBond!( uc::UnitCell , base::Int64 , target::Int64 , offset::Vector{Int64} , mat::Matrix{<:Number} , dist::Float64, label::String )

Add a bond with the given attributes to UnitCell. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond.

TightBindingToolkit.UCell.addBasisSite! — Method

addBasisSite!( uc::UnitCell , position::Vector{Float64} )
addBasisSite!( uc::UnitCell , position::Vector{Float64} , field::Vector{Float64} )

Add a sublattice to the UnitCell at the given real-space position, with an on-site field.

TightBindingToolkit.UCell.addIsotropicBonds! — Method

addIsotropicBonds!( uc::UnitCell , dist::Float64 , mats::Number , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)))
addIsotropicBonds!( uc::UnitCell , dist::Float64 , mats::Matrix{<:Number} , label::String; checkOffsetRange::Int64=1 , subs::Vector{Int64}=collect(1:length(uc.basis)) )

Add a set of "isotropic" bonds, which are the same for each pair of sites at the given distance. If given mat attribute is a number, it is converted into a 1x1 matrix when entered into the bond. The input checkOffsetRange must be adjusted depending on the input distance. The optional input subs is meant for isotropic bonds when only a subset of sublattices are involved.

TightBindingToolkit.UCell.getDistance — Method

getDistance(uc::UnitCell, base::Int64, target::Int64, offset::Vector{Int64}) --> Float64

get the distance between site at position (0, base) and (R, target), where R = offset, when written in units of the unit cell primitive vectors.

TightBindingToolkit.UCell.isSameBond — Method

Function to check if two bond objects are describing the same physical bond, just inverted!

TightBindingToolkit.UCell.isSameUnitCell — Method

isSameUnitCell(uc_1::UnitCell, uc_2::UnitCell) --> Bool

Function to check if two unit cell live on the same underlying lattice or not, i.e. have the same primitives, same sublattices, and same localDim.